[v0.90.2][tests] Split runtime_v2 tests by responsibility (Closes #2308)
