+++ adl/src/runtime_v2/tests/common.rs
@@ -1,0 +1,13 @@
+use std::{
+    env,
+    path::PathBuf,
+    time::{SystemTime, UNIX_EPOCH},
+};
+
+pub(super) fn unique_temp_path(label: &str) -> PathBuf {
+    let nanos = SystemTime::now()
+        .duration_since(UNIX_EPOCH)
+        .expect("time went backwards")
+        .as_nanos();
+    env::temp_dir().join(format!("runtime-v2-{label}-{}-{nanos}", std::process::id()))
+}
